5/5 stage completed

+++ Recipes/task/src/recipes/RecipeController.java
@@ -2,10 +2,13 @@ package recipes;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import recipes.security.User;
+import recipes.security.UserService;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -15,17 +18,27 @@ import java.util.Optional;
 
 @RestController
 @Validated
+@RequestMapping("/api/recipe")
 public class RecipeController {
+
     @Autowired
     RecipeService recipeService;
 
-    @PostMapping("/api/recipe/new")
-    Map<String, Long> setRecipe(@Valid @RequestBody Recipe recipe) {
+    @Autowired
+    UserService userService;
+
+    @PostMapping("/new")
+    Map<String, Long> setRecipe(
+            @Valid @RequestBody Recipe recipe,
+            @AuthenticationPrincipal UserDetails details
+    ) {
+        User user = userService.getUserByEmail(details.getUsername());
+        recipe.setUserCreate(user);
         long id = recipeService.saveRecipe(recipe);
         return Map.of("id", id);
     }
 
-    @GetMapping("/api/recipe/{id}")
+    @GetMapping("/{id}")
     Recipe getRecipe(@PathVariable int id) {
         Optional<Recipe> recipeById = recipeService.getRecipeById(id);
         if (recipeById.isEmpty()) {
@@ -34,7 +47,7 @@ public class RecipeController {
         return recipeById.get();
     }
 
-    @GetMapping(value = "/api/recipe/search")
+    @GetMapping("/search")
     List<Recipe> getCategoryListRecipes
             (
                     @RequestParam(value = "category", required = false) String category,
@@ -49,24 +62,44 @@ public class RecipeController {
         }
     }
 
-    @PutMapping("/api/recipe/{id}")
+    @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    void updateRecipe(@Valid @PathVariable long id, @Valid @RequestBody Recipe recipe) {
+    void updateRecipe
+            (
+                    @Valid @PathVariable long id,
+                    @Valid @RequestBody Recipe recipe,
+                    @AuthenticationPrincipal UserDetails details
+            ) {
         Optional<Recipe> recipeById = recipeService.getRecipeById(id);
         if (recipeById.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
+        String recipeBy = recipeById.get().getUserCreate().getEmail();
+        if (!recipeBy.equals(details.getUsername())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+        User user = userService.getUserByEmail(details.getUsername());
+        recipe.setUserCreate(user);
         recipeService.saveRecipe(recipe, id);
     }
 
-    @DeleteMapping("/api/recipe/{id}")
-    ResponseEntity<?> deleteRecipe(@Valid @PathVariable long id) {
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    void deleteRecipe
+            (
+                    @Valid @PathVariable long id,
+                    @AuthenticationPrincipal UserDetails details
+            ) {
         Optional<Recipe> recipeById = recipeService.getRecipeById(id);
         if (recipeById.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
+        String recipeBy = recipeById.get().getUserCreate().getEmail();
+        if (!recipeBy.equals(details.getUsername())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         recipeService.deleteRecipeById(id);
-        return ResponseEntity.noContent().build();
     }
 }
+
